:wrench: Allow specification of custom timeout for API calls
Kraken API can be a bit shaky at times and the timeout always defaults
to 5000ms. Adapt the `KrakenClient` constructor to take an
`options`-object instead of just the `otp`-parameter which can also
hold a custom timeout value.

Ensure backwards compatibility by allowing options to be a string
still (which will default to just `options.otp` being set to that
string). Timeout still defaults to 5000ms so the change should be 100%
backwards-compatible.